Full coverage for saml + split tests
